+++ backend/src/main/java/com/example/unify/models/ForwardGeocodeResponse.java
@@ -3,7 +3,7 @@ package com.example.unify.models;
 import lombok.Data;
 
 @Data
-public class ForwardGeocodeResponse{
-    private int latitude;
-    private int longitude;
+public class ForwardGeocodeResponse {
+    private double latitude;
+    private double longitude;
 }
